+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 Alfresco JS API
 
-<a name="1.3.1"></a>
+<a name="1.4.0"></a>
 # [1.4.0](https://github.com/Alfresco/alfresco-js-api/releases/tag/1.4.0) (27-04-2017)
 ## Fix
 - [NodePagingList type should include optional "source" property](https://github.com/Alfresco/alfresco-js-api/issues/211)
@@ -12,6 +12,9 @@ Alfresco JS API
 - [fix documentation comments Api](https://github.com/Alfresco/alfresco-js-api/issues/214)
 - [Revert changes added to task-forms rest api](https://github.com/Alfresco/alfresco-js-api/issues/217)
 - [isLoggedIn method of EcmAuth client is returning "true" when ticket expires](https://github.com/Alfresco/alfresco-js-api/issues/210)
+
+## Features
+- [Add the Alfresco Discovery API to the JS-API](https://github.com/Alfresco/alfresco-js-api/issues/197)
 
 <a name="1.3.0"></a>
 # [1.3.0](https://github.com/Alfresco/alfresco-js-api/releases/tag/1.3.0) (05-04-2017)